Adds wallclock based Alarms to fibers (#27)
Adds wallclock based alarms to fibers. These enable for the setting of accurate alarms based on a fixed unix epoch or in the local timezone. These alarms are resistant to NTP updates and the drift of an unreliable monotonic clock. The key functions provided are :wait_next_op(t) and :wait_next(t) that allow for the setting of alarms based on time tables (eg. {wday=2, hour=9} for an alarm that return at 9am on a Monday morning) and :wait_absolute(epoch) and :wait_absolute_op(epoch), that take a Unix epoch number as a float.

This commit also adds tests and examples.